requirements.txt
SQLAlchemy==1.3.11
Flask_Jsonpify==1.5.0
tweepy==3.8.0
Flask==1.1.1
Flask_RESTful==0.3.7
python-dotenv==0.10.3
